2d screen scheme for debug predictor
*  interaction from watcher to server by zerorpc
*  SeenToScreenTranslator and some funcs are now in seen_to_screen.py file
*  BasicPredictor as superclass for GoodPredictor and DebugPredictor
*  mouse movement is now part of predictor classes
