+++ watcher/test.py
@@ -6,7 +6,10 @@ import threading
 import zerorpc
 import numpy as np
 import drawer
-
+import from_internet_or_for_from_internet.PNP_solver as pnp_solver
+import cv2
+import predictor_module
+import zerorpc
 
 def thread_func(debug_predictor, backend_server):
     while True:
@@ -30,7 +33,7 @@ def thread_func(debug_predictor, backend_server):
             pass
 
 
-class DebugPredictor:
+class DebugPredictor(predictor_module.BasicPredictor):
     def __init__(self):
         self.backend = subprocess.Popen(["python", "backend.py"], stdout=subprocess.PIPE)
         frontend = subprocess.Popen(["./frontend/node_modules/.bin/electron", "./frontend"])
@@ -40,15 +43,42 @@ class DebugPredictor:
         self.plane = np.array([0., 1., 0., -100.])
         thread = threading.Thread(target=thread_func, args=(self, self.backend))
         thread.start()
+        self.solver = pnp_solver.PoseEstimator()
         self.client = zerorpc.Client()
         self.client.connect("tcp://127.0.0.1:4242")
 
     def predict_eye_vector_and_face_points(self, imgs, time_now, configurator=None):
-        return [imgs], [self.right_eye_gaze], [self.left_eye_gaze], [np.zeros((68, 3))], {}
+        projections, _ = cv2.projectPoints(self.solver.model_points_68, np.array([0., 0., 0.]), np.array([0., 0., 0.],),
+                                        self.solver.camera_matrix, self.solver.dist_coeefs)
+        projections = projections.reshape((-1, 2))
+        return imgs, [self.right_eye_gaze], [self.left_eye_gaze], [projections], {}
+
+    def get_mouse_coords(self, cameras, time_now):
+        face, results, out_inform = self.predict(cameras, time_now)
+        # out_inform["rotator"] -= angle
+        # out_inform["cutter"] -= offset
+        print(str(out_inform))
+
+        result = list(results)
+        results = [0, 0]
+        for cam in result:
+            for eye in cam:
+                results[0] += eye[0]
+                results[1] += eye[1]
+        results[0] /= len(result) * 2
+        results[1] /= len(result) * 2
+        if results[0] < 0:
+            results[0] = 0
+        if results[1] < 0:
+            results[1] = 0
+        if results[0] >= 1920:
+            results[0] = 1919
+        if results[1] >= 1080:
+            results[1] = 1079
+        self.client.set_mouse_position(*results)
 
 
 if __name__ == "__main__":
     pred = DebugPredictor()
     while True:
-        print(pred.gaze_vector)
         pass
